Rename some files and functions/vars for simplicity.
