+++ src/frame-comparisons.js
@@ -1,21 +1,19 @@
-// integer-frame-comparisons.js
-const { ZERO } = require('./integer-frames.js');
+const { ZERO } = require('./valid-frames.js');
 
 /*
- * isInteger
- * Returns true if the frame represents some integer, false if not.
+ * Returns true if the frame represents some valid frame, false if not.
  * IMPORTANT: We should not expose this directly, but write wrappers for it.
  *
  * @param  {Frame} frame
  *         Frame of OCR data.
- * @param  {IntegerFrame} integerFrame
- *         Valid Frame of OCR data representing our integer. We compare the input frame
+ * @param  {Frame} validFrame
+ *         Valid Frame of OCR data representing some desired character, such as an integer. We compare the input frame
  *         against this valid frame..
  * @return {Boolean}
  *         Returns true if our OCR equals validFrame
  */
-function isInteger(frame, integerFrame) {
-    return JSON.stringify(frame) === JSON.stringify(integerFrame)
+function isValidFrameFor(frame, validFrame) {
+    return JSON.stringify(frame) === JSON.stringify(validFrame)
 }
 
 /*
@@ -27,7 +25,7 @@ function isInteger(frame, integerFrame) {
  *         Returns true if our OCR equals ZERO
  */
 function isZero(frame) {
-    return isInteger(frame, ZERO);
+    return isValidFrameFor(frame, ZERO);
 }
